+++ src/net_permissions.h
@@ -6,6 +6,7 @@
 #define BITCOIN_NET_PERMISSIONS_H
 
 #include <netaddress.h>
+#include <netbase.h>
 
 #include <string>
 #include <type_traits>
@@ -14,6 +15,11 @@
 struct bilingual_str;
 
 extern const std::vector<std::string> NET_PERMISSIONS_DOC;
+
+/** Default for -whitelistrelay. */
+static const bool DEFAULT_WHITELISTRELAY = true;
+/** Default for -whitelistforcerelay. */
+static const bool DEFAULT_WHITELISTFORCERELAY = false;
 
 enum class NetPermissionFlags : uint32_t {
     None = 0,
@@ -89,7 +95,9 @@ public:
 class NetWhitelistPermissions : public NetPermissions {
 public:
     static bool TryParse(const std::string &str,
-                         NetWhitelistPermissions &output, bilingual_str &error);
+                         NetWhitelistPermissions &output,
+                         ConnectionDirection &output_connection_direction,
+                         bilingual_str &error);
     CSubNet m_subnet;
 };