+++ test/functional/p2p_permissions.py
@@ -107,7 +107,21 @@ class P2PPermissionsTests(BitcoinTestFramework):
             ["bypass_proof_request_limits"],
         )
 
+        for flag, permissions in [
+            (["-whitelist=noban,out@127.0.0.1"], ["noban", "download"]),
+            (["-whitelist=noban@127.0.0.1"], []),
+        ]:
+            self.restart_node(0, flag)
+            self.connect_nodes(0, 1)
+            peerinfo = self.nodes[0].getpeerinfo()[0]
+            assert_equal(peerinfo["permissions"], permissions)
+
         self.stop_node(1)
+        self.nodes[1].assert_start_raises_init_error(
+            ["-whitelist=in,out@127.0.0.1"],
+            "Only direction was set, no permissions",
+            match=ErrorMatch.PARTIAL_REGEX,
+        )
         self.nodes[1].assert_start_raises_init_error(
             ["-whitelist=oopsie@127.0.0.1"],
             "Invalid P2P permission",