+++ doc/release-notes.md
@@ -4,4 +4,6 @@ Bitcoin ABC version 0.30.5 is now available from:
 
   <https://download.bitcoinabc.org/0.30.5/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+  - Additional flags "in" and "out" have been added to `-whitelist` to control whether
+    permissions apply to incoming connections and/or manual (default: incoming only).